# cortes-ia-contracts/design/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cortes IA — contrato proposto",
    "version": "0.2.0",
    "description": "Contrato-alvo atualizado pelas decisões v0.2. Não confundir com OpenAPI efetivamente gerado pela API inicial; ver API_IMPLEMENTADA.md."
  },
  "servers": [
    {
      "url": "https://api.example.invalid",
      "description": "Placeholder deliberado; substituir pelo host do ambiente."
    }
  ],
  "paths": {
...
    "/internal/v1/jobs/{jobId}/lease": {
      "post": {
        "operationId": "internal_lease",
        "summary": "Adquirir tentativa exclusiva",
        "description": "Ownership e validação server-side; regras no plano técnico.",
        "tags": [
          "internal"
        ],
        "security": [
          {
            "serviceAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lease"
                }
              }
            }
          },
          "default": {
            "description": "Erro sanitizado; códigos e condições descritos em CONVENCOES.md.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeaseInput"
              }
            }
          }
        }
      }
    },
    "/internal/v1/jobs/{jobId}/heartbeat": {
      "post": {
        "operationId": "internal_heartbeat",
        "summary": "Renovar lease validando fencing",
        "description": "Ownership e validação server-side; regras no plano técnico.",
        "tags": [
          "internal"
        ],
        "security": [
          {
            "serviceAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lease"
                }
              }
            }
          },
          "default": {
            "description": "Erro sanitizado; códigos e condições descritos em CONVENCOES.md.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Heartbeat"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-cortes_session"
      },
      "serviceAuth": {
        "type": "mutualTLS",
        "description": "Workload autenticado no ingresso privado; sem acesso público."
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "code": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "traceId"
        ]
      },
...
      "LeaseInput": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "messageId": {
            "type": "string",
            "format": "uuid"
          },
          "expectedGeneration": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": [
          "messageId",
          "expectedGeneration"
        ]
      },
      "Lease": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "jobId": {
            "type": "string",
            "format": "uuid"
          },
          "attemptId": {
            "type": "string",
            "format": "uuid"
          },
          "leaseToken": {
            "type": "integer",
            "minimum": 1
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "inputManifestId": {
            "type": "string",
            "format": "uuid"
          },
          "disposition": {
            "type": "string",
            "enum": [
              "GRANTED",
              "ALREADY_DONE",
              "BUSY",
              "CANCELLED"
            ]
          }
        },
        "required": [
          "jobId",
          "disposition"
        ]
      },
      "Heartbeat": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "attemptId": {
            "type": "string",
            "format": "uuid"
          },
          "leaseToken": {
            "type": "integer",
            "minimum": 1
          },
          "sequence": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": [
          "attemptId",
          "leaseToken",
          "sequence"
        ]
      },
...
    }
  }
}